Update OAuth and Authentication top level configs
These are pretty aggressive changes:

1. Add AuthenticationType to handle user facing login
2. Remove key configuration on all config and secret references
3. Remove DenyAll IDP (implied by no IDPs configured)
4. Remove bool config from AllowAll (type is sufficient)
5. Order all IDPs alphabetically
6. Remove grantMethod config (it was incorrectly placed inside of
the IDP config when it is global for the OAuth server).  It is
never configured by users because they have no idea what it does.
Without a strong understanding of our OAuth implementation, the docs
are not helpful:
https://docs.okd.io/latest/install_config/configuring_authentication.html#grant-options
https://docs.okd.io/latest/architecture/additional_concepts/authentication.html#oauth-clients
I have also never seen anyone change the default set by ansible:
https://github.com/openshift/openshift-ansible/blob/master/roles/openshift_control_plane/defaults/main.yml#L135
7. Remove the generate mappingMethod.  I have never seen this used
and it simply does not line up with how RBAC/group sync/etc work.
8. Change name of keystone deprecated field to useUsernameIdentity
9. Make ca required for request header.  Any cluster running without
a ca set is broken, and should just use AllowAll.
10. Reordered OAuth related fields to be consistent across IDPs
11. Remove legacy field from GitLab, OIDC is required now.

Signed-off-by: Monis Khan <mkhan@redhat.com>